+++ src/index.ts
@@ -1,4 +1,5 @@
 export { FlightClient } from './flight-client';
+export { resolveDriverLib, driverPlatform, cachedDriverPath, driverVersion } from './driver-lib';
 export { FlightSQLClient } from './flightsql-client';
 export * from './types';
 export * from './errors';